Fix issue with tab not switching content between history and data
