+++ raven_front/lib/pages/wallet/asset_details/asset_details_bloc.dart
@@ -1,18 +1,9 @@
 import 'dart:async';
 
-import '../../../utils/data.dart';
-import 'package:flutter/material.dart';
 import 'package:raven_back/raven_back.dart';
 import 'package:raven_back/services/transaction/transaction.dart';
-import 'package:raven_front/pages/wallet/asset_details/assset_details_components.dart';
 import 'package:raven_front/services/lookup.dart';
-import 'package:raven_front/services/storage.dart';
-import 'package:raven_front/utils/data.dart';
-import 'package:raven_front/utils/extensions.dart';
-import 'package:raven_front/widgets/widgets.dart';
-import 'package:raven_front/components/components.dart';
 import 'package:rxdart/rxdart.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 AssetDetailsBloc get assetDetailsBloc => AssetDetailsBloc.instance();
 
@@ -31,7 +22,7 @@ class AssetDetailsBloc {
     }));
     listeners.add(streams.client.busy.listen((bool value) {
       if (!value) {
-       // setState(() {});
+        // setState(() {});
       }
     }));
   }
