--- raven_front/lib/pages/wallet/asset_details/asset_details_datasource.dart
@@ -1,3 +1,0 @@
-class AssetDetailsDataSource{
-  
-}
